Bring motorcar qt dependency up to date with qtwayland v5.4
This is a first draft at the process.  As of now I am not able to get this to detect the Oculus dk1.  Not sure if this is a problem with the oculus sdk, dk1 hardware or the changes this commit introduces.
QWayland Surfaces are split into a model view pattern
QWayland Surfaces require a QWaylandBufferAttacher
Some previous private headers are no longer used.
The example QWayland compositors have changed the way they handle setup that may be benificial to follow in motorcar.
The final commit of this branch will remove all my nast @@JAF comment sections notating the sections of change introduced.  Seemed easier then going through commit logs.